Addressed PR notes from @boxhock
